+++ common/src/main/java/earth/terrarium/cadmus/common/commands/ModCommands.java
@@ -4,6 +4,7 @@ import com.mojang.brigadier.CommandDispatcher;
 import earth.terrarium.cadmus.common.commands.claims.ClaimCommand;
 import earth.terrarium.cadmus.common.commands.claims.ClaimInfoCommand;
 import earth.terrarium.cadmus.common.commands.claims.UnclaimCommand;
+import earth.terrarium.cadmus.common.commands.claims.admin.AdminClaimCommands;
 import earth.terrarium.cadmus.common.commands.claims.admin.AdminCommands;
 import earth.terrarium.cadmus.common.commands.claims.admin.AdminFlagCommands;
 import net.minecraft.commands.CommandBuildContext;
@@ -18,6 +19,7 @@ public class ModCommands {
         UnclaimCommand.register(dispatcher);
         ClaimInfoCommand.register(dispatcher);
         AdminCommands.register(dispatcher);
+        AdminClaimCommands.register(dispatcher);
         AdminFlagCommands.register(dispatcher);
     }
 }