+++ common/src/main/java/earth/terrarium/cadmus/common/commands/claims/UnclaimCommand.java
@@ -10,6 +10,7 @@ import earth.terrarium.cadmus.common.util.ModUtils;
 import net.minecraft.commands.CommandSourceStack;
 import net.minecraft.commands.Commands;
 import net.minecraft.commands.arguments.coordinates.ColumnPosArgument;
+import net.minecraft.network.chat.Component;
 import net.minecraft.server.level.ColumnPos;
 import net.minecraft.server.level.ServerPlayer;
 import net.minecraft.world.level.ChunkPos;
@@ -20,6 +21,12 @@ public class UnclaimCommand {
 
     public static void register(CommandDispatcher<CommandSourceStack> dispatcher) {
         dispatcher.register(Commands.literal("unclaim")
+            .then(Commands.literal("all")
+                .executes(context -> {
+                    ServerPlayer player = context.getSource().getPlayerOrException();
+                    CommandHelper.runAction(() -> clear(player));
+                    return 1;
+                }))
             .then(Commands.argument("pos", ColumnPosArgument.columnPos())
                 .executes(context -> {
                     ServerPlayer player = context.getSource().getPlayerOrException();
@@ -36,14 +43,17 @@ public class UnclaimCommand {
 
     public static void unclaim(ServerPlayer player, ChunkPos pos) throws ClaimException {
         Pair<String, ClaimType> claimData = ClaimHandler.getClaim(player.serverLevel(), pos);
-        if (claimData == null) {
-            throw ClaimException.CHUNK_NOT_CLAIMED;
-        }
+        if (claimData == null) throw ClaimException.CHUNK_NOT_CLAIMED;
         boolean isMember = TeamHelper.isMember(claimData.getFirst(), player.server, player.getUUID());
-        if (!isMember) {
-            throw ClaimException.DONT_OWN_CHUNK;
-        }
+        if (!isMember) throw ClaimException.DONT_OWN_CHUNK;
         ModUtils.tryClaim(player.serverLevel(), player, Map.of(), Map.of(pos, ClaimType.CLAIMED));
         player.displayClientMessage(CommonUtils.serverTranslatable("text.cadmus.unclaiming.unclaimed_chunk_at", pos.x, pos.z), false);
+    }
+
+    public static void clear(ServerPlayer player) {
+        String id = TeamHelper.getTeamId(player.getServer(), player.getUUID());
+        player.server.getAllLevels().forEach(l -> ClaimHandler.clear(l, id));
+        Component name = TeamHelper.getTeamName(id, player.server);
+        player.displayClientMessage(CommonUtils.serverTranslatable("text.cadmus.clear", name == null ? player.getDisplayName().getString() : name.getString()), false);
     }
 }