+++ changelog.md
@@ -1,5 +1,27 @@
+# -----{ 0.0.6 }-----
+- Add new admin commands for claiming and unclaiming others' claims.
+- Add admin list command to display all teams and their respective IDs.
+- Add `/claim clear` command to clear all claimed chunks.
+- Add an REI Favorite Entry for opening the claim map.
+- Move all admin claiming related commands to `/cadmus adminclaims`
+
+# -----{ 0.0.5 }-----
+- Fix Forge Server Crash.
+
 # -----{ 0.0.4 }-----
 
-Fix entity griefing crash (#9).
-Fix chunk loading not working when no players are online (#10).
-Fix not being optional on client on Forge.
+- Fix entity griefing crash (#9).
+- Fix chunk loading not working when no players are online (#10).
+- Fix not being optional on client on Forge.
+
+# -----{ 0.0.2 }-----
+- Update to 1.20.1 Fix #7
+
+# -----{ 0.0.1 }-----
+- Add JourneyMap Compat
+- Add unclaim command for Admin claims and fix admin remove command
+- Change admin commands from "/claim admin" to "/cadmus admin"
+- Fix water buckets breaking on Forge
+
+# -----{ 0.0.0 }-----
+- Initial Release.